moves top level namespaces into their own files
